// src/app/generated-code/model/user.ts
/**
 * car_rental_backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';
import { Status } from './status';


export interface User { 
    id?: number;
    name?: string;
    lastName?: string;
    email?: string;
    phone?: string;
    username?: string;
    password?: string;
    googleLogin?: boolean;
    status?: Status;
    darkMode?: boolean;
    imageUrl?: string;
    imagePublicId?: string;
    role?: Role;
    createdAt?: string;
    lastUpdate?: string;
    businessId?: number;
}